Begin isolating filter unit tests
